actualizacion

+++ Grupal2/Assets/Roseling/BaseUnits.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public abstract class BaseUnits : MonoBehaviour, IDamageable
 {
-    public int lifeMax = 100f;
-    public int lifeAct;
-    public int damage = 10f;
+    public float lifeMax = 100f;
+    public float lifeAct;
+    public float damage = 10f;
 
     protected virtual void Start()
     {
@@ -15,7 +15,7 @@ public abstract class BaseUnits : MonoBehaviour, IDamageable
     {
         lifeAct -= damage;
 
-        if (lifeAct <= 0)
+        if (lifeAct <= 0f)
         {
             Dead();
         }
